backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
#psycopg2-binary==2.9.7
psycopg[binary]==3.2.10
django-extensions==3.2.3

# Authentication & Security
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0
django-ratelimit==4.1.0

# API Integrations
requests==2.31.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
tweepy==4.14.0
linkedin-api==2.2.0

# Analytics & Data Processing
pandas==2.2.3
numpy==1.26.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
django-plotly-dash

# Task Queue & Background Jobs
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# File Handling & Media
Pillow==10.4.0
django-storages==1.14.2
boto3==1.34.0

# Email & Notifications
django-anymail==10.1
sendgrid==6.10.0
# django-email-backends==0.1.0

# Search & Filtering
django-filter==23.3
django-haystack==3.2.1
elasticsearch==8.11.0

# Monitoring & Debugging
django-debug-toolbar==4.2.0
sentry-sdk==1.38.0
django-silk==5.0.4

# Development Tools
black==23.11.0
flake8==6.1.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
django-redis==5.4.0

# Additional Utilities
python-decouple==3.8
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-widget-tweaks==1.5.0
django-import-export==3.3.1
